requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
aiosqlite>=0.18.0
aiofiles==23.1.0
aio-pika==9.5.7
pdfminer.six
python-multipart==0.0.6
pydantic
pytest==7.4.0
httpx==0.24.1
